Support large Api responses
Pivoted to stream responses from the api call instead of fetching it whole and converting it to my ApiResponse class. Without it, we'd reach an in-memory buffer limit.

Also, changed the Feature class to achieve this as we'd need to only parse the feature array of the response. This makes the ApiResponse and Property classes obsolete tho